move semantic nullability tests to their own file
